requirements.txt
aiohttp
pyyaml
requests
logging
asyncio
gitpython
